Refactor: Move ElmInlineText and ElmInlineIcon components to typography directory
- Updated import paths for ElmInlineText and ElmInlineIcon in various components and stories to reflect their new location in the typography directory.
- Created new ElmInlineIcon and ElmInlineText components with appropriate props and styles.
- Adjusted related story files to ensure proper functionality and display of the new components.
