+++ packages/core/src/components/badge/ElmTag.vue
@@ -19,7 +19,7 @@
 
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 import { getLuminance } from 'polished'
 
 export interface ElmTagProps {
