+++ packages/core/src/components/containments/ElmSnackbar.vue
@@ -15,7 +15,7 @@
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
 import { VNode } from 'vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 
 export interface ElmSnackbarProps {
   label?: string
