+++ packages/core/src/components/containments/ElmToggle.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script setup lang="ts">
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 import { Icon } from '@iconify/vue'
 import type { Property } from 'csstype'
