+++ packages/core/src/components/data/ElmStatusMessage.vue
@@ -40,7 +40,7 @@
 
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 
 export interface ElmStatusMessageProps {
   status: 'success' | 'error' | 'warning' | 'pending'
