+++ packages/core/src/components/icon/ElmLoginIcon.vue
@@ -27,7 +27,7 @@
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
 import ElmTooltip from '../containments/ElmTooltip.vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 
 export interface ElmLoginIconProps {
   /**
