+++ packages/core/src/components/jsonschema/ElmFieldAttribute.vue
@@ -10,7 +10,7 @@
 
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 
 export interface ElmFieldAttributeProps {
   icon: string
