+++ packages/core/src/components/media/ElmFile.vue
@@ -33,7 +33,7 @@
 
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 import type { Property } from 'csstype'
 
 export interface ElmFileProps {
