+++ packages/core/src/components/navigation/ElmBookmark.vue
@@ -58,7 +58,7 @@
 
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 import ElmImage from '../media/ElmImage.vue'
 import type { Property } from 'csstype'
