+++ packages/core/src/components/navigation/ElmTableOfContents.vue
@@ -21,7 +21,7 @@
 
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from '../typography/ElmInlineText.vue'
 
 export interface ElmTableOfContentsProps {
   headings: Array<{
