+++ packages/core/src/components/typography/ElmBulletedList.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/vue3'
 import ElmBulletedList from './ElmBulletedList.vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from './ElmInlineText.vue'
 
 const meta: Meta<typeof ElmBulletedList> = {
   title: 'Components/Typography/ElmBulletedList',
