+++ packages/core/src/components/typography/ElmHeading.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/vue3'
 import ElmHeading from './ElmHeading.vue'
-import ElmInlineText from '../inline/ElmInlineText.vue'
+import ElmInlineText from './ElmInlineText.vue'
 
 const meta: Meta<typeof ElmHeading> = {
   title: 'Components/Headings/ElmHeading',
