+++ packages/core/src/components/typography/ElmInlineIcon.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/vue3'
 import ElmInlineIcon from './ElmInlineIcon.vue'
 
 const meta: Meta<typeof ElmInlineIcon> = {
-  title: 'Components/Inline/ElmInlineIcon',
+  title: 'Components/Typography/ElmInlineIcon',
   component: ElmInlineIcon,
   tags: ['autodocs'],
   args: {}
