+++ packages/core/src/components/typography/ElmInlineIcon.vue
