+++ packages/core/src/components/typography/ElmInlineText.vue
