+++ packages/core/src/components/typography/ElmTypingAnimation.vue
